Mise à jour Dragoman — ajout convocations (distinct), factures, paiements, inventaire, AD status, tolklink, génération PDF, améliorations UI/CSS, .gitignore nettoyage bin/obj/.vs

+++ Dragoman.Server/Models/Indexation.cs
@@ -5,7 +5,7 @@ namespace Dragoman.Server.Models;
 
 public partial class Indexation
 {
-    public byte IdIndex { get; set; }
+    public int IdIndex { get; set; }
 
     public DateTime Startdate { get; set; }
